some more rearranging
This refactors all the CREATE TABLE and INSERT needed for setup, previously all grouped into the db_schema, into files specifically grouped for the resource type.

Next step is moving the resource-specific queries (both the statement preparation and the record scanning) into these same files and out of db.go ... it may be worthwhile to lean on sqlx for the object mapping instead of writing all the type-matching and struct-traversing code out.  Main benefit would be only having to depend on the DB driver and few other things.  For the moment, the surface area of the table types is still somewhat small and it's not going to change.

But sqlx or gorm will be very useful for the wits-go repo.