generic Table/Record/Column abstractions, additional method requirements on Record interface, clarity on Table interface
everything typechecks but there are several // TODO markers where functionality needs to be filled in, remaining work is mechanical and then the transition to a reflection-based CreateSql is possible (but the question remains whether  it's worth it -- probably not for this repo but valuable takeaways here nonetheless)

will try Ent in the next project.. generated code can be just as type-safe as generic code (or more, in some cases) but may not allow for as much flexibility (like here, being able to iterate over a []TableSql slice for creating and initializing the database schema) or expressiveness of having the DB schema as tags on the in-memory (and JSON) representation.  Ent seems to require some familiarity with which code is generated and which golang files to refer to for source/editing.  But it's less work than rolling my own ORM!